fix: ignore setsid error when pid is same as pgid
When running jailer with --daemonize, the jailer used to fail
to start if it was started a process group leader, because setsid()
returns an error if the calling process is already a process group
leader.
This was a common reported issue by users running jailer manually.
Compared to maintaining a document which was rather difficult to
understand, it is easier to handle this case in jailer by ignoring
setsid failure when process id is same as process groud id.

Signed-off-by: Sudan Landge <sudanl@amazon.com>